backend/requirements.txt
fastapi
uvicorn
requests
gtfs-realtime-bindings
python-dotenv
redis